Fixed projectiles being unable to damage Damageables becauseit lookedfor Damageable<float>

+++ Dungeon Scramblers/Assets/Scripts/Loadout/Equippables/ProjectileStats.cs
@@ -100,18 +100,17 @@ public class ProjectileStats : MonoBehaviour
     // For Abilities object to collide, the opposing object must have a 2D collider as well as a Rigidbody2D
     protected virtual void OnTriggerEnter2D(Collider2D collision)
     {
-
-        if(collision.tag == "Shield")
+        
+        if (collision.tag == "Shield")
         {
             ResetProjectiles();
             return;
         }
 
         //get Damageable from collision object
-        IDamageable<float> damageable = collision.GetComponent<IDamageable<float>>();
+        IDamageable<int> damageable = collision.GetComponent<IDamageable<int>>();
         //if the collision does not exist or damageable is not on the object
         if (collision == null || damageable == null)  return;
-
 
         //If the colliding object inherits from Abstract player then apply the status effect(s)
         AbstractPlayer unit = collision.gameObject.GetComponent<AbstractPlayer>();
